+++ .gitignore
@@ -19,4 +19,4 @@ typings/
 jsconfig.json
 
 # Generated by Webpack
-www/js/dist
+# www/js/
