+++ www/index.html
@@ -155,6 +155,6 @@
 		</div>
 		<script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
 		<script src="cordova.js"></script>
-		<script type="module" src="./js/dist/build.js"></script>
+		<script type="module" src="./js/index.js"></script>
 	</body>
 </html>
